refactor: /profile divide Profile component to smaller component & add tabs
