+++ src/components/Profile/components/MainInfo/components/AboutMe/AboutMe.module.css
@@ -1,0 +1,24 @@
+.about {
+    word-break: break-word;
+    padding: 10px 5px;
+}
+
+.about p {
+    font-family: 'Roboto';
+    font-weight: 300;
+    font-size: 16px;
+    margin: 25px 0px;
+}
+
+.about b {
+    font-family: 'Roboto';
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 28px;
+}
+
+.noData {
+    font-family: 'Roboto';
+    font-weight: 700;
+    font-size: 24px;
+}
