+++ src/components/Profile/components/MainInfo/components/AboutMe/index.js
@@ -1,0 +1,1 @@
+export { AboutMe } from './AboutMe';
