+++ src/components/Profile/components/MainInfo/components/Proofs/Proofs.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Proofs = () => {
+  return (
+    <div>Proofs</div>
+  )
+}
