+++ src/components/Profile/components/MainInfo/components/Proofs/index.js
@@ -1,0 +1,1 @@
+export { Proofs } from './Proofs';
