+++ src/components/Profile/components/MainInfo/index.js
@@ -1,0 +1,1 @@
+export { MainInfo } from './MainInfo';
