+++ src/components/Profile/components/UserInfo/ProfileInfo/ProfileInfo.module.css
@@ -4,6 +4,7 @@
     flex-direction: row;
     align-items: center;
     padding: 25px;
+    padding-left: 50px;
 }
 
 .itemText {
@@ -12,9 +13,6 @@
     margin-left: 20px;
     word-break: break-word;
 }
-
-
-
 
 .infoItem svg {
     font-size: xx-large;
