+++ src/components/Profile/components/UserInfo/UserInfo.module.css
@@ -1,0 +1,5 @@
+.info {
+    display: flex;
+    flex-wrap: wrap;
+    flex-basis: 40%;
+}
