+++ src/components/Profile/components/UserInfo/index.js
@@ -1,0 +1,1 @@
+export { UserInfo } from './UserInfo';
